qcacld-3.0: Add support for new Aggressive roaming mode
When userspace sends new Aggressive roaming mode,
host should be able to store the aggressive roam config
INI values.

Add support in roaming to handle new Aggressive roaming
mode.

CRs-Fixed: 3847154

+++ components/mlme/dispatcher/inc/wlan_mlme_public_struct.h
@@ -2257,7 +2257,7 @@ struct wlan_mlme_lfr_cfg {
 	bool disconnect_on_nud_roam_invoke_fail;
 	bool hs20_btm_offload_disable;
 	uint32_t roam_aggre_scan_step_rssi;
-	uint32_t roam_aggre_threshold;
+	uint8_t roam_aggre_threshold;
 };
 
 /**
